[refactor] BOM 캐시 로직 고아까지 조회 추가
