+++ smerp/src/main/java/com/domino/smerp/bom/repository/cache/BomCostCacheQueryRepositoryImpl.java
@@ -1,4 +1,4 @@
-package com.domino.smerp.bom.repository.query;
+package com.domino.smerp.bom.repository.cache;
 
 
 import com.domino.smerp.bom.dto.response.BomRequirementResponse;
