+++ smerp/src/main/java/com/domino/smerp/bom/repository/query/BomQueryRepository.java
@@ -1,5 +1,10 @@
 package com.domino.smerp.bom.repository.query;
 
+import java.util.List;
+
 public interface BomQueryRepository {
+
+  // 고아 데이터까지 찾아오기
+  List<Long> findAllBomAndOrphanItemIds();
 
 }
